refactor: BaseballGame Class를 기능에 따라서 BaseballGameController, BaseballGameModel, Eenemy, HintSentenceGenerator, Player로 분리
상태를 유지하지고있지 않는 Class에 대해서 싱글톤으로 구현
