+++ src/main/java/baseball/Application.java
@@ -3,7 +3,7 @@ package baseball;
 public class Application {
 	public static void main(String[] args) {
 		//TODO: 숫자 야구 게임 구현
-		BaseballGame baseballGame = new BaseballGame();
+		BaseballGameController baseballGame = new BaseballGameController();
 		do {
 			baseballGame.startGame();
 		} while (baseballGame.decideToRestart());
